Move TaskGroup out to separate file in src/Query/
